third_party/blink/web_tests/platform/generic/svg/dynamic-updates/SVGFEMergeNodeElement-svgdom-in-prop-expected.txt

SVG 1.1 dynamic update tests

Tests dynamic updates of the 'in' property of the SVGFEMergeNodeElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS mergeNode1.in1.baseVal is "SourceGraphic"
PASS mergeNode1.in1.baseVal is "off"
PASS successfullyParsed is true

TEST COMPLETE
